premiers test

+++ src/mcp.h
@@ -52,13 +52,13 @@ class mcp {
         void begin();
         void setupPortA(uint8_t polarity, uint8_t pullup, uint8_t dir);
         void setupPortB(uint8_t polarity, uint8_t pullup, uint8_t dir);
-        byte readGPIO(byte port);
-        void writeGPIO(byte port, byte value);
+        byte readGPIO(uint8_t port);
+        void writeGPIO(uint8_t port, uint8_t value);
         void writePin(uint8_t pin, bool value);
         bool readPin(uint8_t pin);
     private:
         uint8_t _chipAddr;
-        void mcu_write(int registre, int value);
+        void mcu_write(uint8_t registre, uint8_t value);
 };
 
 #endif
